third_party/blink/web_tests/fast/dom/navigation-type-reload-expected.txt

This tests that performance.navigation.type is 1

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigation.type is 1
PASS successfullyParsed is true

TEST COMPLETE
